Rename AxE related variables
Rename the variables related to the MC in the macro for the mass fits, and related to the AxE in all the committed files, to take into account that the AxE is different for different processes. The AxE implemented for now is the one related to coherent j/psi photoproduction.